# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Talenta Submitter API",
    "version": "1.0.0",
    "description": "Automated bulk submission of Talenta timesheets with support for holidays and annual leave."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Base API path"
    }
  ],
  "components": {
    "securitySchemes": {
      "tokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Token",
        "description": "Paste your Talenta session cookie here (e.g., PHPSESSID=your_value)"
      }
    }
  },
  "security": [
    {
      "tokenAuth": []
    }
  ],
  "paths": {
    "/talenta/bulk-submit": {
      "post": {
        "tags": ["Talenta Service"],
        "summary": "Bulk submit Talenta timesheet entries",
        "description": "Submits attendance tasks between the given dates. Automatically skips weekends and uses special task IDs for holidays and annual leave.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fromDate": {
                    "type": "string",
                    "example": "2025-10-02"
                  },
                  "toDate": {
                    "type": "string",
                    "example": "2025-10-02"
                  },
                  "startTime": {
                    "type": "string",
                    "example": "09:00:00"
                  },
                  "activity": {
                    "type": "string",
                    "example": "Meeting with client"
                  },
                  "endTime": {
                    "type": "string",
                    "example": "17:00:00"
                  },
                  "taskId": {
                    "type": "integer",
                    "example": 127520,
                    "description": "Example: Maintenance = 127520, Medistra = 492289"
                  },
                  "holidays": {
                    "type": "array",
                    "items": { "type": "string" },
                    "example": ["2025-10-01"]
                  },
                  "annualLeave": {
                    "type": "array",
                    "items": { "type": "string" },
                    "example": ["2025-10-01"]
                  }
                },
                "required": [
                  "fromDate",
                  "toDate",
                  "startTime",
                  "endTime",
                  "taskId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Bulk submission result (per day status)" },
          "400": { "description": "Missing cookie or invalid body parameters" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/talenta/get-task-list": {
      "get": {
        "tags": ["Talenta Service"],
        "summary": "Get Talenta Task List",
        "description": "Retrieves the list of tasks from Talenta.",
        "responses": {
          "200": { "description": "List of tasks retrieved successfully" },
          "400": { "description": "Missing cookie or invalid body parameters" },
          "500": { "description": "Internal server error" }
        }
      }
    }
  }
}